Generate Open API spec for network attachments
In order to generate the openapi spec json of network attachment
definition, two steps are required:

1. Generate GetOpenAPIDefinitions.go by running:
   ./hack/update-openapigen.sh

2. Generate openapi spec file by running:
   go run ./hack/gen-openapi-spec/main.go v1

Due to lack of support for CRD as specified on [1], modifying vendored
file was required.

The alternative is replacing this patch with the output of:
`kubectl get --raw /openapi/v2` from a running cluster.

[1] https://github.com/kubernetes/kube-openapi/issues/174#issuecomment-541520186